update base card search
